requirements.txt
asgiref==3.5.2
Django==4.1
django-ckeditor==6.5.0
django-filter==22.1
django-grappelli==3.0.3
django-js-asset==2.0.0
djangorestframework==3.13.1
mysqlclient==2.1.1
pytz==2022.2.1
sqlparse==0.4.2
